fix(AutoChoice): rounds must be Int — Double.random reintroduces drift
The round-6 floor-divide anchor relies on (currentLap - rounds) * 360
being an exact integer multiple of 360. When rounds was Double.random
(5.0...8.0), a fractional rounds value (e.g. 6.7) made the result
fractional-multiples-of-360 + target, leaving a sub-360° residue and
reintroducing the same modular drift the fix was meant to remove.

Caught by CI run 25674830649 — testSpinRotationLandsAtTargetModulo360
asserted final mod 360 == target mod 360 and got off-by-tens-of-degrees
errors. Switching to Int.random(in: 5...8) restores the invariant.

Visual rotation behavior unchanged — Int rounds = 5/6/7/8 are all > 4
full turns CCW which is plenty for the spin animation feel.

+++ AutoChoice/Models/WheelStore.swift
@@ -108,10 +108,14 @@ final class WheelStore {
         // rotation drifts and the segment under the pointer no longer matches `chosen`.
         // Fix: anchor the new rotation to a multiple-of-360 baseline derived from the
         // current rotation (via floor division), then add the deterministic target.
+        // Note: `rounds` must be an Int (whole turns) so that
+        //   (currentLap - rounds) * 360 + target
+        // is exactly an integer multiple of 360 plus target — fractional rounds would
+        // leave a sub-360° residue and reintroduce the drift the fix is meant to remove.
         let target = -(Double(idx) * segment + segment / 2)
-        let rounds = Double.random(in: 5...8)
+        let rounds = Int.random(in: 5...8)
         let currentLap = (currentRotation / 360).rounded(.down)
-        currentRotation = (currentLap - rounds) * 360 + target
+        currentRotation = (currentLap - Double(rounds)) * 360 + target
         lastResult = chosen
         history.insert(HistoryEntry(listName: list.name, choice: chosen.label, timestamp: .now), at: 0)
         let cap = isPremium ? Self.historyCap : Self.freeHistoryCap
